Review.R: 1. To avoid errors when time series is all NA, require at least 100 measurements for SmoothInterp. 2. Make loop through plot functions skip cases where there is no named function matching index
VarList:
Added DBARU to be able to plot it with DBARP in RPlot16

RPlot16.R:
added DBARU to middle panel of plot

RPlot19:
changed limits on ad. pot. T to ones more appropriate for WINTER

RPlot22:
changed to plot if either UHSAS or PCASP meet test; previously, required both, which prevented plot when either was missing.